Update WorkingTest.py
here a change has been made in the starting function
previously it created object no matter the round and of same name, that is it rewrote the object and hence the pool data was re written too

but now the previous object is passed as an argument which if present will pass on the pool info to the new object.
this will allow the other attributes of the object to be remade so that they can perform their function appropriately where as if the data would have been re written while over writing the other attributes too then there would have been too much lines

so now new objects are createrd for every round so that the data is reset
but if it's not the first round then the pool will be copied into the new objct and other attributes will remain new in new object.
this will help in continuation of the game

this will also delete the old object after copying the data
if it is present in the first place..