Update gameplay_layout testing.py
now this has every aspect or part of the gameplay layout layed in except for cards

now cards will be designed and made
then this funtion will be modified to display cards one by one for both the dealers and players side

a  function will be made to handle this action
in it function will be called whenever the player has chosen an option, this will here be emulated as a buttom press
then there will be a an incrementing equation to update the coordinates for the cards to be printed at
there will be such equation for the dealer also,
so this wil print the cards
these incremented coordinates will be variables outside this function that will be passed in as arguments
also, the values needed for cards to display wil be passed as arguments too.

the main display for cards will be inside this function only
this function will be called when we need to display cards
first the coordinates will be calculated by this function then the  card printing function will be called for both the player and dealer.

but all this is later
now need to work on cards