Update gameplay_layout testing.py
so this function is now able to print cards on both dealer and the player side
and in proper incrementation too
now there are 2 different functions to be called each for a player and dealer card

so now the basic gameplay layout has been set
and now has to think about  how will it be laid inside the display file

because the display functions are inside this file so will have to think of something and all.
bsides this
the card row for the splitting and the horizontal card for the double down is yet to be designed so will add them later.
those will mark the end  of display part of the game

then the things will be left are
sequence managing by calling functions and debugging process