Update gameplay_layout testing.py
so in this file now i am trying the different approaches to make this file work as function.
so the thing is  that the functions can be called inside to get the images and are able to  well make a big backgroung image that just needs to be printed but the thing is the image cannot be called on in other outside functions (the functions that i made to print the cards) without passsing them as arguments but it would be better to have the background creating funtions all in one place so placed these tqo functions inside one function.

now will test the input and output functionlity of function wil video format and all that stuff
so that i can output when i want