Update gameplay_layout testing.py
so this is working but the cards are getting on only for a split second because the frrame is remade everytime the function is called so thinking a work around
thinking of usig the balck image globally and making changes to it permanently about the stuff that neeeds to be there for long.
but not the menu or stats as they require to be printed continuously so
planning on emptying that portion of the image. at the end of the while loop.