Update WorkingTest.py
corrections and modifications to have only uniques cards to be used for the game and no card shal be repeated. and the required changes so this functionality wil lbe reset at the start of every round so that

also this here file shows changes done to factorize the existence of another object after splitting so that game will work normally even when the cards are split.