Create Composite_Functions_File.py
this file was to made the testing of the things easier for me while designing the sequence and stuff of this game.

this has all the displaying and working functions
this is imported to the menu interaction file to work with the process

now this file leads to changes and debugging and then these changes are made in the original display and working functions file. because those files will be used to make the final file that will eventually be made into the .exe file.
and also that these 2 files are to help in observing the required functins for the working of the task and the changes can be done to them while keeping eye on the corresponding working or the displaying function required in a a task