Update Composite_Functions_File.py
this is to improve on the winning condition check while using split function

this is now going to return different values for each object if the cardvalues are equal accordingly.

and the else is added for whent the spllit is not used because if it wasn't there, then the function will check the condition twice and will give problems.