Update Composite_Functions_File.py
these are the changes that ended up edited due to split action of the game testing
there were many errors too but now they have been corrected but the working of splitting is not good so will remove it
from the game
will remove the splitting action from the game completely
well atleast the visible part
backend is too much complicated to remove some feature now
honestly i don't have energy to do so
i might not even remove splitting now
most probably no
but still uk will think about this in the morning.
but for now will work on the options 5 in the gamplay layout then will correct that 'a' error that i have been having
then will work on end menu options then in the end will work on the starting menu options

then after all this i will ive final touches
mostly in the main file of the game then will turn it to .exe then will share for approval and suggestions.