Update Composite_Functions_File.py
so now am trying solve the 'a' problem
the problem is when a is the latest card
the index i gets back to the index of value popped that is it resets the index position for proper functioning.
but this still leaves a positon by the loop to visit where there is no value after pooping or even an index
this is bevause loop was told to got to the value before the popping occured so that end condition won't change

so that's why the error is showing that the index is out of range
because there is no value there now.
so will add a few lines of if conditin to the card 'a' statement
this has been understood by testing the working of pop() function and the postiton of pointer as it works
i testes different loops and then i observed that there ewas that error here too
l = [0,9,8,7]
>>> for i in range(len(l)):
...     print(i)
...     l.pop()
...     print(l[i])

this one is the loop that had the error
at the end of loop
because here also
pointer is at the second last position according to the loop because of the pop statement as the last value has been popped
but there is no index or value over there so index gets out of range