Update Composite_Functions_File.py
so just kept it simple
i didn't do anything to the value just left it as it is
as '0' would make no change in the sum and jsut put the flag = 1
not the value of a is calculated and added to the sum and correct value is displayed.
now the problem is added that if a black jack is dealt in the first round only
the bet amount keeps increasing till it reaches infinty so will work on this first
this will be solved by calling the bet increment at the end menu
and not in the black jack check function.