Update Composite_Functions_File.py
in pool value function a parameter has been added to check the value of pool so that if the pool value is zero and the round is not first one then the normal operations won't happen and only pool value will be zero, this will rpevent the previous error of negative pool value in case of intermidieate round and valuebeing zero

also, a little calculation error in black jack value was there, idon't know why it was so just turned the value to be 1.5 time the bet
that's jsut same as before but calculation is different

also the betting error was also corrected that when the bets larger thatn the pool were accepted
this was because of comparing sign being wrong, it was to check whether bet was smaller but it should have checked for it being larger.

also dd check and splitting checks are to be chagned for these changes in pool value function. as now the pool value 0 will also allow for next round and won't check the number.